Move most conversion logic out of index.ts
